packages: SystemUI: improve the VPN footer layout
Not only is the background color of the VPN footer layout a different color
but the text is not centered. The padding was just off.

- Also cloned the styles for this layout and made the text bold. Google clearly
wants the user to know that they're connected to a VPN by adding a statusbar
icon and giving this its own footer. This just adds to that by making it stick
out slightly.

Change-Id: Ia7a0df63d4a4b6e583235323561eeab34e83065b
Signed-off-by: Anirudh Gupta <anirudhgupta109@gmail.com>
Signed-off-by: Diab Neiroukh <officiallazerl0rd@gmail.com>
Signed-off-by: spezi77 <spezi7713@gmx.net>