Update date on keyguard when ticking past midnight
When the time passes 23:59 to 00:00 (midnight) the date on keyguard does
not always change. ACTION_DATE_CHANGED only triggers on a change from
the user (or perhaps timezone etc). Instead we should also include
ACTION_TIME_TICK in the clock update filter to ensure that the date is
updated correctly when changing days.

Test: none performed

Signed-off-by: Joshua Hesketh <josh@nitrotech.org>

Change-Id: Ib7e514a9107d40a3eb0100a80b66307f9e471e00
Signed-off-by: spezi77 <spezi7713@gmx.net>